requirements.txt
automium==0.2.6
Click==7.0
Django==2.2.6
django-cors-headers==3.1.1
django-filter==2.2.0
djangorestframework==3.10.3
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
legacy==0.1.6
Markdown==3.1.1
MarkupSafe==1.1.1
pkg-resources==0.0.0
pytz==2019.3
sqlparse==0.3.0
Werkzeug==0.16.0
